Fix warnings about size_t to uint truncation
Use size_t for the internal hashing in QTextFormat to be
in line with QHash.

+++ src/gui/text/qtextformat_p.h
@@ -96,7 +96,7 @@ public:
 
     FormatVector formats;
     QVector<qint32> objFormats;
-    QMultiHash<uint,int> hashes;
+    QMultiHash<size_t,int> hashes;
 
     inline QFont defaultFont() const { return defaultFnt; }
     void setDefaultFont(const QFont &f);
